Feature #305 Check and implement prototype for Hadoop integration
First implementation of Bloob (Hadoop Blob Store). Not working, yet, due to Weld issues.

+++ server/server/core.api/src/main/java/com/puresoltechnologies/purifinity/server/core/api/analysis/store/FileInformation.java
@@ -10,25 +10,25 @@ import com.puresoltechnologies.commons.misc.hash.HashId;
  */
 public class FileInformation {
 
-	private final HashId hashId;
-	private final int size;
+    private final HashId hashId;
+    private final long size;
 
-	public FileInformation(HashId hashId, int size) {
-		super();
-		this.hashId = hashId;
-		this.size = size;
-	}
+    public FileInformation(HashId hashId, long size) {
+	super();
+	this.hashId = hashId;
+	this.size = size;
+    }
 
-	public HashId getHashId() {
-		return hashId;
-	}
+    public HashId getHashId() {
+	return hashId;
+    }
 
-	public int getSize() {
-		return size;
-	}
+    public long getSize() {
+	return size;
+    }
 
-	@Override
-	public String toString() {
-		return hashId + "(size: " + size + " bytes)";
-	}
+    @Override
+    public String toString() {
+	return hashId + "(size: " + size + " bytes)";
+    }
 }
